+++ src/yakc/systems/z1013.cc
@@ -162,7 +162,7 @@ z1013::step_debug() {
     }
     while ((old_pc == cpu.PC) && !cpu.INV);    
     this->decode_video();
-    return all_ticks;
+    return uint32_t(all_ticks);
 }
 
 //------------------------------------------------------------------------------
